Move epoll socket tests into the sockets_node suite
